refactor: move sts under credentials
