+++ internal/credentials/fetch.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/aripalo/vegas-credentials/internal/sts"
+	"github.com/aripalo/vegas-credentials/internal/credentials/sts"
 
 	"github.com/dgraph-io/badger/v3"
 	"github.com/dustin/go-humanize"
